requirements.txt
scrapy>=2.11,<3.0
beautifulsoup4>=4.12
lxml>=5.0
pyyaml>=6.0
click>=8.1
pytest>=8.0
